Feature/t1242 voting

+++ contracts/voting/VotingTimelock.sol
@@ -103,7 +103,7 @@ contract VotingTimelock is NamedContract, VotingTimelockStorage, VotingTimelockE
         uint256 value,
         string calldata signature,
         bytes calldata data,
-        uint32 eta
+        uint256 eta
     ) external returns (bytes32) {
         require(
             msg.sender == _guardian,
